v1.0.0

+++ esbox/models/paddle_model.py
@@ -58,7 +58,7 @@ class PaddleModel(nn.Layer):
         i = 0
         arrays = []
         for shape in array_shapes:
-            size = np.prod(shape, dtype=np.int)
+            size = np.prod(shape, dtype=np.int32)
             array = flat_array[i:(i + size)].reshape(shape)
             arrays.append(array)
             i += size
